mapping first

+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/ControleurRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Controleur;
 
-public interface ControleurRepo extends JpaRepository<Controleur, Integer> {
+public interface ControleurRepo extends JpaRepository<Controleur, Long> {
 }
