+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/EntiteRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Entite;
 
-public interface EntiteRepo extends JpaRepository<Entite, Integer> {
+public interface EntiteRepo extends JpaRepository<Entite, Long> {
 }
