+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/EquipementRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Equipement;
 
-public interface EquipementRepo extends JpaRepository<Equipement, Integer> {
+public interface EquipementRepo extends JpaRepository<Equipement, Long> {
 }
