+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/FormulaireRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Formulaire;
 
-public interface FormulaireRepo extends JpaRepository<Formulaire, Integer> {
+public interface FormulaireRepo extends JpaRepository<Formulaire, Long> {
 }
