+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/JustificationRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Justification;
 
-public interface JustificationRepo extends JpaRepository<Justification, Integer> {
+public interface JustificationRepo extends JpaRepository<Justification, Long> {
 }
