+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/QuestionRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Question;
 
-public interface QuestionRepo extends JpaRepository<Question, Integer> {
+public interface QuestionRepo extends JpaRepository<Question, Long> {
 }
