+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/ResponsableControleurRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.ResponsableControleur;
 
-public interface ResponsableControleurRepo extends JpaRepository<ResponsableControleur, Integer> {
+public interface ResponsableControleurRepo extends JpaRepository<ResponsableControleur, Long> {
 }
