+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/ResponsableGeneralRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.ResponsableGeneral;
 
-public interface ResponsableGeneralRepo extends JpaRepository<ResponsableGeneral, Integer> {
+public interface ResponsableGeneralRepo extends JpaRepository<ResponsableGeneral, Long> {
 }
