+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/SectionRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Section;
 
-public interface SectionRepo extends JpaRepository<Section, Integer> {
+public interface SectionRepo extends JpaRepository<Section, Long> {
 }
