+++ src/main/java/org/system/digitalisationservicedecontrole/repositories/UniteRepo.java
@@ -3,5 +3,5 @@ package org.system.digitalisationservicedecontrole.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.system.digitalisationservicedecontrole.entities.Unite;
 
-public interface UniteRepo extends JpaRepository<Unite, Integer> {
+public interface UniteRepo extends JpaRepository<Unite, Long> {
 }
